deletw unused firestore methods. moved to backend
